Decoupling caching from AsyncMultiMap.
